(#18) Correcao de erro chave primaria

+++ drogaria/src/main/java/br/com/drogaria/domain/Categoria.java
@@ -18,7 +18,7 @@ public class Categoria {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Byte codigo;
+	private Short codigo;
 	
 	@Column(length = 50, nullable = false, unique = true)
 	private String nome;
